savepoint - refactor auth for pairing
